requirements.txt
django
django-debug-toolbar==2.2
gunicorn==20.0.4
psycopg2-binary==2.9.1
gunicorn
python-dotenv
Pillow
djangorestframework
markdown
celery
django-celery-beat
twilio
djangorestframework-simplejwt
drfaddons
drf_yasg
pyyaml
smsaero_api